requirements.txt
mysql-connector-python
SQLAlchemy==2.0.1
sqlalchemy-utils
fastapi[all]~=0.89.1
fastapi-utils
PYMYSQL~=1.0.2
uvicorn~=0.20.0
pydantic~=1.10.4
cryptography
starlette~=0.22.0
fastapi-mail
pytest
coverage
PyJWT==2.6.0
python-decouple==3.7
bcrypt